remove duplicate deposit class
